requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Auth
python-jose[cryptography]>=3.3.0

# MongoDB
motor>=3.3.0

# Evaluation
jiwer>=3.0.0

# Templates
jinja2>=3.1.0

# ML Backend (Apple Silicon)
mlx>=0.5.0; sys_platform == "darwin" and platform_machine == "arm64"
mlx-lm>=0.12.0; sys_platform == "darwin" and platform_machine == "arm64"

# ML Backend (Linux/CUDA) - uncomment if needed
# transformers>=4.40.0
# peft>=0.10.0
# torch>=2.2.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
